requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
idna==3.4
kafka-python==2.0.2
pydantic==2.5.1
pydantic-core==2.14.3
pymongo==4.6.0
python-dateutil==2.8.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.8.0
uvicorn==0.24.0.post1
